Refactor pattern.py
