+++ onnxscript/rewriter/pattern.py
@@ -2,2188 +2,41 @@
 # Licensed under the MIT License.
 from __future__ import annotations
 
-import abc
-import contextlib
-import dataclasses
-import enum
-import inspect
-import itertools
-import math
-from collections import defaultdict
-from collections.abc import Mapping
-from typing import (
-    Any,
-    Callable,
-    Iterable,
-    Iterator,
-    MutableSequence,
-    Protocol,
-    Sequence,
-    TypeVar,
-    Union,
+from onnxscript.ir import _tape
+from onnxscript.rewriter._basics import MatchingTracer, MatchResult, MatchStatus
+from onnxscript.rewriter._matcher import PatternMatcher, SimplePatternMatcher
+from onnxscript.rewriter._pattern_ir import (
+    ANY_VALUE,
+    Constant,
+    OpsetPatternBuilder,
+    OrValue,
+    pattern_builder,
 )
-
-import onnxscript.optimizer
-from onnxscript import ir
-from onnxscript.ir import _convenience, _tape
-
-T = TypeVar("T")
-
-
-class Pattern(Protocol[T]):  # type: ignore[misc]
-    """This is essentially a Predicate[T], that is, a Callable[[T], bool] bound to the name "matches"."""
-
-    def matches(self, item: T) -> bool: ...
-
-
-class StringPattern(abc.ABC, Pattern[str]):
-    """Abstract base class for string patterns."""
-
-    @abc.abstractmethod
-    def matches(self, item: str) -> bool:
-        pass
-
-    @abc.abstractmethod
-    def __str__(self) -> str:
-        pass
-
-
-class StringConstantPattern(StringPattern):
-    """Matches strings with given value."""
-
-    def __init__(self, value: str):
-        self._value = value
-
-    def matches(self, item: str) -> bool:
-        return item == self._value
-
-    def __str__(self) -> str:
-        return self._value
-
-    def value(self) -> str:
-        return self._value
-
-
-class PrefixPattern(StringPattern):
-    """Matches strings with a given prefix."""
-
-    def __init__(self, value: str) -> None:
-        self._value = value
-
-    def matches(self, value: str) -> bool:
-        return value.startswith(self._value)
-
-    def __str__(self) -> str:
-        return f"{self._value}*"
-
-
-class AttrPattern(Pattern[Union[ir.Attr, ir.RefAttr]]):
-    """Base class for an attribute pattern. Matches any attribute value by default."""
-
-    def __init__(self, name: str | None):
-        self._name = name
-
-    @property
-    def name(self) -> str | None:
-        return self._name
-
-    def matches(self, attr: ir.Attr | ir.RefAttr) -> bool:
-        return True
-
-    def __str__(self) -> str:
-        return self._name if self._name is not None else "anonymous:" + str(id(self))
-
-
-# TODO: Support tensors. Align with usage elsewhere.
-SupportedAttrTypes = Union[
-    int,
-    float,
-    str,
-    Sequence[int],
-    Sequence[float],
-    Sequence[str],
-]
-
-
-class AttrConstantPattern(AttrPattern):
-    """Matches attributes with given value.
-
-    Uses standard equality for matching. For list-valued attributes, the order of elements matters.
-    If order is immaterial, we need to define a separate pattern for that.
-    """
-
-    def __init__(self, value: SupportedAttrTypes):
-        super().__init__(None)
-        self._value = value
-
-    def matches(self, attr: ir.Attr | ir.RefAttr) -> bool:
-        return isinstance(attr, ir.Attr) and attr.value == self._value
-
-    def __str__(self) -> str:
-        return str(self._value)
-
-
-def _to_attr_pattern(value: AttrPattern | ValuePattern | SupportedAttrTypes) -> AttrPattern:
-    """Represents promotion of values allowed as keyword-arguments in a pattern-builder call to an AttrPattern."""
-    if isinstance(value, AttrPattern):
-        return value
-    if type(value) is ValuePattern:
-        # This is a hack. Currently, when we create pattern-variables, we create them as ValuePattern,
-        # and change them to AttrPattern if/when used in an attribute context. We could use type
-        # annotations to distinguish between ValuePattern and AttrPattern, but forces users to
-        # use these type annotations.
-        # TODO: check for misuse at rule-creation time. (Currently will be caught by matcher at match-time.)
-        return AttrPattern(value.name)
-    if isinstance(value, (int, float, str)):
-        return AttrConstantPattern(value)
-    if isinstance(value, Sequence):
-        if all(isinstance(i, (int, float)) for i in value):
-            return AttrConstantPattern(value)
-        if all(isinstance(i, str) for i in value):
-            return AttrConstantPattern(value)
-        raise ValueError("Only lists of int/float/str can be used as an AttrPattern")
-    raise TypeError(f"Cannot convert {type(value)} to AttrPattern")
-
-
-class OpsetPatternBuilder:
-    """Represents an opset pattern and a pattern builder.
-
-    (i) It is used to create a NodePattern (via OpPatternBuilder).
-    Example usage:
-    ::
-
-        z = op.Matmul(x, y)
-
-    Here, `op` is an instance of OpsetPatternBuilder and `op.Matmul` is an instance
-    of OpPatternBuilder, and  `op.Matmul(x, y)` is an instance of NodePattern.
-
-    (ii) It contains a domain pattern matched against the actual opset domain used in the
-    input model.
-    """
-
-    def __init__(self, domain: StringPattern | str, record: bool = False) -> None:
-        if isinstance(domain, str):
-            domain = StringConstantPattern(domain)
-        self._domain_pattern = domain
-        if record:
-            self._nodes: list[NodePattern] | None = []
-        else:
-            self._nodes = None
-
-    def domain_pattern(self) -> StringPattern:
-        return self._domain_pattern
-
-    def __getattr__(self, op_name: str) -> OpPatternBuilder:
-        return OpPatternBuilder(self, op_name)
-
-    def submodule(self, name: str) -> OpPatternBuilder:
-        """This method is used to match against submodule ops with prefix."""
-        return OpPatternBuilder(self, PrefixPattern(name))
-
-    def __str__(self) -> str:
-        return str(self._domain_pattern)
-
-    def add_node(self, node: NodePattern) -> None:
-        if self._nodes is not None:
-            self._nodes.append(node)
-
-    def nodes(self) -> Sequence[NodePattern]:
-        if self._nodes is None:
-            raise ValueError("Nodes were not recorded.")
-        return self._nodes
-
-
-onnxop = OpsetPatternBuilder("")
-
-torch_module_op = OpsetPatternBuilder(PrefixPattern("pkg.torch"))
-
-
-class OpPatternBuilder:
-    """A utility class to build a NodePattern.
-
-    It is used primarily to create a NodePattern.
-    Example usage:
-    ::
-
-        z = op.Matmul(x, y)
-
-    Here, `op` is an instance of OpsetPatternBuilder and `op.Matmul` is an instance
-    of OpPatternBuilder, and  `op.Matmul(x, y)` is an instance of NodePattern.
-
-    """
-
-    def __init__(
-        self,
-        pattern_builder: OpsetPatternBuilder,
-        op_name: str | Pattern[str],
-    ) -> None:
-        self.pattern_builder = pattern_builder
-        self.op_name = op_name
-
-    def __call__(
-        self,
-        *args,
-        _domain: str | None = None,
-        _version: int | None = None,
-        _outputs: int | list[str | None] = 1,
-        _allow_other_attributes: bool | None = None,
-        _allow_other_inputs: bool | None = None,
-        **kwargs,
-    ):
-        if _version is not None:
-            raise ValueError(
-                "The pattern builder does not support '_version' keyword argument. "
-                "Version restrictions should be handled by rewrite rules."
-            )
-        if _domain is None:
-            opset_pattern = self.pattern_builder.domain_pattern()
-        elif isinstance(_domain, str):
-            opset_pattern = StringConstantPattern(_domain)
-        else:
-            # TODO(rama): allow OpsetPatternBuilder as _domain.
-            raise TypeError("_domain must be a string.")
-
-        if isinstance(_outputs, int):
-            _outputs = [None for _ in range(_outputs)]
-        elif not isinstance(_outputs, Sequence) or not all(
-            isinstance(x, (str, type(None))) for x in _outputs
-        ):
-            raise ValueError("_outputs must be an int or a list[str|None].")
-        inputs = [_to_value_pattern(x) for x in args]
-        attributes = {name: _to_attr_pattern(value) for (name, value) in kwargs.items()}
-        node_pattern = NodePattern(
-            opset_pattern,
-            self.op_name,
-            inputs,
-            attributes,
-            _outputs,
-            allow_other_attributes=_allow_other_attributes,
-            allow_other_inputs=_allow_other_inputs,
-        )
-        self.pattern_builder.add_node(node_pattern)
-        output_values = node_pattern.outputs
-        # Unpack outputs if there is only one output, the common case.
-        if len(output_values) == 1:
-            return output_values[0]
-        else:
-            return output_values
-
-
-def _to_value_pattern(
-    x: ValuePattern | int | float | None,
-) -> ValuePattern | None:
-    """Promotes an input-value used to construct a NodePattern to a ValuePattern.
-
-    Example usage:
-    ::
-        x = op.MatMul(a, b)
-        z = op.Add(x, 0)
-
-    In this example, `a, `b`, and `x` are ValuePatterns used to construct a NodePattern.
-    `0` is a constant (int) value, and is automatically promoted to a ValuePattern.
-
-    Note that this is a shorthand for creating a Constant pattern. The user can more
-    explicitly write this as:
-    ::
-        z = op.Add(x, op.Constant(0))
-    """
-    if x is None or isinstance(x, ValuePattern):
-        return x
-    if isinstance(x, (int, float)):
-        return Constant(x)
-    if isinstance(x, Sequence):
-        if all(isinstance(i, (int, float)) for i in x):
-            return Constant(x)
-        raise ValueError("Only lists of int/float can be used as a ValuePattern")
-
-    raise TypeError(f"Cannot convert {type(x)} to ValuePattern")
-
-
-class MatchResult:
-    """The state object used by the pattern-matching algorithm.
-
-    A match can either succeed or fail.
-    If it succeeds, it returns a list of nodes that matched the pattern
-    and a set of bindings for the variables in the pattern.
-
-    Example:
-    ::
-        def pattern(x, shape1, shape2):
-            t1 = op.Reshape(x, shape1)
-            t2 = op.Reshape(t1, shape2)
-            return t2
-    The above pattern matches a sequence of two Reshape ops.
-    The matched_nodes will contain the two Reshape ops, and the bindings will
-    contain the values that are bound to the variables `x`, `shape1`, and `shape2`.
-    """
-
-    def __init__(self) -> None:
-        # We use a stack of partial matches to handle OR patterns that require backtracking.
-        self._partial_matches: list[PartialMatchResult] = [PartialMatchResult()]
-
-    @property
-    def _current_match(self) -> PartialMatchResult:
-        """Returns the current match result."""
-        return self._partial_matches[-1]
-
-    def enter_new_match(self) -> None:
-        """Starts a new sub-match to try out one of multiple alternatives."""
-        match = PartialMatchResult()
-        self._partial_matches.append(match)
-
-    def abandon_current_match(self) -> PartialMatchResult:
-        """Abandons the current alternative due to failure."""
-        if len(self._partial_matches) < 2:
-            raise ValueError("No match to abandon.")
-        return self._partial_matches.pop()
-
-    def merge_current_match(self) -> None:
-        """Merges a successful sub-match for an alternative with the parent one."""
-        if len(self._partial_matches) < 2:
-            raise ValueError("No match to merge.")
-        current_match = self._partial_matches.pop()
-        previous_match = self._partial_matches[-1]
-        if not current_match:
-            raise ValueError("Current match is not successful.")
-        # Merge the two matches.
-        previous_match.merge(current_match)
-
-    def __bool__(self) -> bool:
-        """Returns True if the current match is successful."""
-        return bool(self._current_match)
-
-    def fail(
-        self,
-        reason: str = "",
-        failure_source: Union[ir.Node, ir.Value, list[Union[ir.Node, ir.Value]]] | None = None,
-    ) -> MatchResult:
-        self._current_match.fail(reason, failure_source)
-        return self
-
-    @property
-    def reason(self) -> str:
-        """Returns the reason for the failure."""
-        return self._current_match.reason
-
-    @property
-    def nodes(self) -> Sequence[ir.Node]:
-        """Returns the list of nodes that matched the pattern."""
-        return self._current_match.nodes
-
-    def bind_node(self, pattern_node: NodePattern, node: ir.Node):
-        """Binds a pattern node to a matched node."""
-        self.add_node(node)
-        self._current_match.node_bindings[pattern_node] = node
-
-    def add_node(self, node: ir.Node) -> None:
-        """Adds a node to the list of matched nodes."""
-        self._current_match.add_node(node)
-
-    def bind_value(self, pattern_value: ValuePattern, value: Any) -> bool:
-        var_name = pattern_value.name
-        # TODO(rama): Simplify the following. We currently bind values to
-        # pattern variables in two different ways: via their name, or via the
-        # pattern-value itself.
-        if var_name is None:
-            for match in self._partial_matches:
-                if pattern_value in match.value_bindings:
-                    # TODO(rama): Use appropriate equality-check here.
-                    if match.value_bindings[pattern_value] == value:
-                        return True
-                    self._current_match.fail(
-                        f"Binding failure: {pattern_value} bound to two different values.",
-                        [match.value_bindings[pattern_value], value],
-                    )
-                    return False
-            self._current_match.value_bindings[pattern_value] = value
-            return True
-        return self.bind(var_name, value)
-
-    def bind(self, var: str, value: Any) -> bool:
-        for match in self._partial_matches:
-            if var in match.bindings:
-                # TODO(rama): Use appropriate equality-check here.
-                if match.bindings[var] == value:
-                    return True
-                self._current_match.fail(
-                    f"Binding failure: {var} bound to two different values.",
-                    [match.bindings[var], value],
-                )
-                return False
-        self._current_match.bindings[var] = value
-        return True
-
-    @property
-    def bindings(self) -> dict[str, Any]:
-        """Returns the bindings for the pattern variables."""
-        if len(self._partial_matches) > 1:
-            raise ValueError("Bindings can be accessed only at the top-level match.")
-        return self._current_match.bindings
-
-    @property
-    def value_bindings(self) -> dict[ValuePattern, ir.Value]:
-        """Returns the bindings for the value variables."""
-        if len(self._partial_matches) > 1:
-            raise ValueError("Value bindings can be accessed only at the top-level match.")
-        return self._current_match.value_bindings
-
-    @property
-    def outputs(self) -> MutableSequence[ir.Value]:
-        """Returns the list of output values that matched the pattern."""
-        if len(self._partial_matches) > 1:
-            raise ValueError("Outputs can be accessed only at the top-level match.")
-        return self._current_match.outputs
-
-    @property
-    def failure_nodes_and_values(self) -> list[Union[ir.Node, ir.Value]]:
-        """Returns the nodes and values that caused the failure."""
-        return self._current_match._failure_nodes_and_values
-
-    def lookup_node(self, pattern_node: NodePattern) -> ir.Node | None:
-        """Looks up the node that matched the given pattern node."""
-        for match in self._partial_matches:
-            if pattern_node in match.node_bindings:
-                return match.node_bindings[pattern_node]
-        return None
-
-    def num_matched_nodes(self) -> int:
-        """Returns the number of nodes matched so far."""
-        return sum(len(match.node_bindings) for match in self._partial_matches)
-
-
-class PartialMatchResult:
-    """The state object used by the pattern-matching algorithm for a sub-match."""
-
-    def __init__(self) -> None:
-        self._success: bool = True
-        # For a successful match, _matched_nodes is a list of values that matched the pattern.
-        # These include the internal nodes of the pattern that were matched, but not
-        # the leaves (sub-trees) that match against the variables in the pattern.
-        # These represent the values that will be replaced by the replacement pattern.
-        self._matched_nodes: MutableSequence[ir.Node] = []
-        # For a successful match, bindings is a dictionary of mapping pattern-variable-names
-        # to values.
-        self._bindings: dict[str, Any] = {}
-        self._value_bindings: dict[ValuePattern, ir.Value] = {}
-        self._node_bindings: dict[NodePattern, ir.Node] = {}
-
-        self._outputs: list[ir.Value] = []
-        # For a failed match, _reason is a string that describes the reason for the failure.
-        self._reason: str = ""
-        # Track the node(s) or value(s) that caused the failure.
-        self._failure_nodes_and_values: list[Union[ir.Node, ir.Value]] = []
-
-    def __bool__(self):
-        return self._success
-
-    def fail(
-        self,
-        reason: str = "",
-        failure_source: Union[ir.Node, ir.Value, list[Union[ir.Node, ir.Value]]] | None = None,
-    ) -> None:
-        self._success = False
-        self._reason = reason
-        if failure_source is not None:
-            if isinstance(failure_source, list):
-                self._failure_nodes_and_values.extend(failure_source)
-            else:
-                self._failure_nodes_and_values.append(failure_source)
-
-    @property
-    def reason(self) -> str:
-        return self._reason
-
-    @property
-    def nodes(self) -> Sequence[ir.Node]:
-        return tuple(self._matched_nodes)
-
-    def add_node(self, node: ir.Node) -> None:
-        """Adds a node to the list of matched nodes."""
-        self._matched_nodes.append(node)
-
-    @property
-    def bindings(self) -> dict[str, Any]:
-        return self._bindings
-
-    @property
-    def value_bindings(self) -> dict[ValuePattern, ir.Value]:
-        return self._value_bindings
-
-    @property
-    def outputs(self) -> MutableSequence[ir.Value]:
-        return self._outputs
-
-    @property
-    def node_bindings(self) -> dict[NodePattern, ir.Node]:
-        return self._node_bindings
-
-    def merge(self, other: PartialMatchResult) -> None:
-        """Merges a successful sub-match for an alternative with the parent one."""
-        if self._success and other._success:
-            # Merge the two successful matches. Matching algorithm responsible for ensuring
-            # that the two matches are compatible. No need to check for conflicts here.
-            self._bindings.update(other._bindings)
-            self._matched_nodes.extend(other.nodes)
-            # Note: outputs should be set only at end of the (top-level) match. There
-            # should be no outputs in the sub-match.
-            assert not other._outputs
-        else:
-            # This should not happen currently.
-            raise NotImplementedError("Merging failed matches is not yet supported.")
-
-
-_pattern_builder: OpsetPatternBuilder = onnxop
-
-
-@contextlib.contextmanager
-def pattern_builder(builder: OpsetPatternBuilder):
-    global _pattern_builder
-    prev_builder = _pattern_builder
-    _pattern_builder = builder
-    yield
-    _pattern_builder = prev_builder
-
-
-class ValuePattern:
-    """Base class for all patterns that match against IR values.
-
-    This is used primarily to provide operator overloadings for arithmetic
-    operations, so that we can write patterns like `x + 1` and `1 + x`.
-    """
-
-    def __init__(self, name: str | None) -> None:
-        self._name = name
-        # Note: uses will be computed only when the full graph-pattern is constructed.
-        self._uses: list[tuple[NodePattern, int]] = []
-
-    def clone(self, node_map: dict[NodePattern, NodePattern]) -> ValuePattern:
-        del node_map
-        return ValuePattern(self._name)
-
-    @property
-    def name(self) -> str | None:
-        return self._name
-
-    def producer(self) -> NodePattern | None:
-        return None
-
-    def uses(self) -> Sequence[tuple[NodePattern, int]]:
-        return self._uses
-
-    def append_use(self, node: NodePattern, index: int):
-        self._uses.append((node, index))
-
-    def __repr__(self) -> str:
-        return f"ValuePattern({self._name!r})"
-
-    def __add__(self, other):
-        return _pattern_builder.Add(self, other)
-
-    def __radd__(self, other):
-        return _pattern_builder.Add(other, self)
-
-    def __sub__(self, other):
-        return _pattern_builder.Sub(self, other)
-
-    def __rsub__(self, other):
-        return _pattern_builder.Sub(other, self)
-
-    def __mul__(self, other):
-        return _pattern_builder.Mul(self, other)
-
-    def __rmul__(self, other):
-        return _pattern_builder.Mul(other, self)
-
-    def __truediv__(self, other):
-        return _pattern_builder.Div(self, other)
-
-    def __rtruediv__(self, other):
-        return _pattern_builder.Div(other, self)
-
-    def __pow__(self, other):
-        return _pattern_builder.Pow(self, other)
-
-    def __str__(self) -> str:
-        return self._name if self._name is not None else "anonymous:" + str(id(self))
-
-
-class NodePattern:
-    """Represents a pattern that matches against a Node.
-
-    This differs from a NodeOutputPattern in that it matches against a node (which
-    may produce 1 or more outputs), whereas a NodeOutputPattern matches against
-    a specific output of a node.
-
-    Args:
-        domain: pattern to match against the domain of the node.
-        op: pattern or string constant to match against the op_type of the node.
-        inputs: sequence of ValuePatterns (or constants) to match against the inputs of the node.
-        attributes: dictionary of attribute patterns to match against the attributes of the node.
-        outputs: specifies pattern-variable-name for outputs (or None)
-        allow_other_attributes: specifies whether other attributes (not mentioned in `attributes`)
-          are allowed in the node.
-    """
-
-    def __init__(
-        self,
-        domain: StringPattern,
-        op: str | Pattern[str],
-        inputs: Sequence[int | float | ValuePattern | None],
-        attributes: dict[str, AttrPattern],
-        outputs: Sequence[str | None],
-        *,
-        allow_other_attributes: bool | None,
-        allow_other_inputs: bool | None,
-    ):
-        if allow_other_attributes is None:
-            # Default behavior: allow other unmatched attributes in the node.
-            allow_other_attributes = True
-        if allow_other_inputs is None:
-            # TODO(rama): Should we default to True? For now, we preserve the current behavior.
-            allow_other_inputs = False
-        self.domain = domain
-        self.op = StringConstantPattern(op) if isinstance(op, str) else op
-        self.inputs = [_to_value_pattern(x) for x in inputs]
-        self.attributes = attributes
-        self.allow_other_attributes = allow_other_attributes
-        self.allow_other_inputs = allow_other_inputs
-        # In the common case, domain and op are constants, which can be used to optimize matching.
-        if isinstance(op, str) and isinstance(domain, StringConstantPattern):
-            # TODO(rama): support overloaded operators.
-            overload = ""
-            self._op_identifier: ir.OperatorIdentifier | None = (
-                domain.value(),
-                op,
-                overload,
-            )
-        else:
-            self._op_identifier = None
-        self.outputs = [NodeOutputPattern(self, i, name) for i, name in enumerate(outputs)]
-
-        # Update uses for inputs.
-        for index, value in enumerate(self.inputs):
-            if value is not None:
-                value.append_use(self, index)
-
-    def __str__(self) -> str:
-        inputs = ", ".join(str(v) for v in self.inputs)
-        outputs = ", ".join(str(v) for v in self.outputs)
-        attributes = ", ".join(f"{k}={v}" for k, v in self.attributes.items())
-        op = str(self.op)
-        domain = str(self.domain)
-        qualified_op = f"{domain}.{op}" if domain else op
-        inputs_and_attributes = f"{inputs}, {attributes}" if attributes else inputs
-        return f"{outputs} = {qualified_op} ({inputs_and_attributes})"
-
-    def op_identifier(self) -> ir.OperatorIdentifier | None:
-        return self._op_identifier
-
-    @property
-    def op_type(self) -> str:
-        return str(self.op)
-
-    def matches(self, node: ir.Node, match: MatchResult) -> MatchResult:
-        """Matches the pattern represented by self against a node.
-
-        This is purely a local node-level match, and does not consider the subgraph rooted at the node.
-        We check the domain, op_type, and attributes of the node, but not the inputs.
-        """
-        # TODO(rama): Ensure we handle "" and "onnx.ai" correctly.
-        if not self.op.matches(node.op_type):
-            return match.fail(
-                f"OpType mismatch: expected {self.op}, got {node.op_type}.", node
-            )
-        if not self.domain.matches(node.domain):
-            return match.fail(
-                f"Domain mismatch: expected {self.domain}, got {node.domain}.", node
-            )
-
-        for name, attr_pattern in self.attributes.items():
-            attr_value = node.attributes.get(name)
-            if attr_value is None:
-                return match.fail(f"Attribute {name} not found in node.", node)
-            if not attr_pattern.matches(attr_value):
-                return match.fail(
-                    f"Attribute {name} mismatch: expected {attr_pattern}, got {attr_value}.",
-                    node,
-                )
-            if attr_pattern.name is not None:
-                if not match.bind(attr_pattern.name, attr_value):
-                    return match
-
-        if not self.allow_other_attributes:
-            for name in node.attributes:
-                # TODO: Support matching default nodes for attributes.
-                if name not in self.attributes:
-                    return match.fail(f"Attribute {name} not expected in node.", node)
-
-        return match
-
-    def clone(self, node_map: dict[NodePattern, NodePattern], swap: bool) -> NodePattern:
-        inputs = [(v.clone(node_map) if v is not None else None) for v in self.inputs]
-        if swap:
-            assert len(inputs) == 2, (
-                "Internal error: commutative swap applies only to binary ops."
-            )
-            inputs = [inputs[1], inputs[0]]
-        outputs = [value.name for value in self.outputs]
-        copied = NodePattern(
-            self.domain,
-            self.op,
-            inputs,
-            self.attributes,
-            outputs,
-            allow_other_attributes=self.allow_other_attributes,
-            allow_other_inputs=self.allow_other_inputs,
-        )
-        node_map[self] = copied
-        return copied
-
-
-class NodeOutputPattern(ValuePattern):
-    """Represents a pattern that matches against a specific output of a Node.
-
-    This is the primary pattern used to match against computed values, that
-    is values computed using a specific op.
-    """
-
-    def __init__(
-        self, producer: NodePattern, output_index: int, name: str | None = None
-    ) -> None:
-        super().__init__(name)
-        self._producer = producer
-        self._output_index = output_index
-
-    def clone(self, node_map: dict[NodePattern, NodePattern]) -> NodeOutputPattern:
-        return node_map[self._producer].outputs[self._output_index]
-        # return NodeOutputPattern(node_map[self._producer], self._output_index, self._name)
-
-    @property
-    def output_index(self) -> int:
-        return self._output_index
-
-    def producer(self) -> NodePattern:
-        return self._producer
-
-
-Var = ValuePattern
-
-
-class AnyValue(ValuePattern):
-    """Represents a pattern that matches against any value."""
-
-    def __init__(self) -> None:
-        super().__init__(None)
-
-    def clone(self, node_map: dict[NodePattern, NodePattern]) -> AnyValue:
-        # A single instance of AnyValue suffices.
-        return self
-
-
-ANY_VALUE = AnyValue()
-
-
-class Constant(ValuePattern):
-    """Represents a pattern that matches against a scalar constant value."""
-
-    def __init__(
-        self,
-        value: int | float | Sequence[int] | Sequence[float],
-        rel_tol: float = 1e-5,
-        abs_tol: float = 1e-8,
-    ) -> None:
-        super().__init__(None)
-        self._value = list(value) if isinstance(value, Sequence) else value
-        self._rel_tol = rel_tol
-        self._abs_tol = abs_tol
-
-    def clone(self, node_map: dict[NodePattern, NodePattern]) -> Constant:
-        del node_map
-        return Constant(self._value, self._rel_tol, self._abs_tol)
-
-    @property
-    def value(self) -> int | float | list[int] | list[float]:
-        return self._value
-
-    def __str__(self) -> str:
-        return str(self._value)
-
-
-class _OpIdDispatchOr(ValuePattern):
-    """Represents a (restricted) form of value pattern disjunction that enables deterministic matching."""
-
-    def __init__(
-        self,
-        op_to_pattern: Mapping[ir.OperatorIdentifier, tuple[Any, ValuePattern]],
-        name: str | None = None,
-        tag_var: str | None = None,
-    ) -> None:
-        """
-        Initialize an _OpIdDispatchOr pattern.
-
-        Args:
-            op_to_pattern: A dictionary mapping operator identifiers to tuples of tag values and patterns.
-                The keys are operator identifiers, and the values are tuples containing a tag value
-                and a pattern to match against.
-            name: An optional variable name for the pattern. Defaults to None. If present,
-                this name will be bound to the value matched by the pattern.
-            tag_var: An optional variable name for the tag. Defaults to None. If present,
-                it will be bound to a value indicating which alternative was matched.
-        """
-        super().__init__(name)
-        self._op_to_pattern = op_to_pattern
-        self._tag_var = tag_var
-
-    @property
-    def tag_var(self) -> str | None:
-        """Returns the tag variable associated with the OrValue pattern."""
-        return self._tag_var
-
-    def clone(self, node_map: dict[NodePattern, NodePattern]) -> _OpIdDispatchOr:
-        return _OpIdDispatchOr(
-            {k: (v[0], v[1].clone(node_map)) for k, v in self._op_to_pattern.items()},
-            self.name,
-            self._tag_var,
-        )
-
-    def get_pattern(self, value: ir.Value) -> tuple[Any, ValuePattern] | None:
-        """Returns the pattern that should be tried for the given value."""
-        producer = value.producer()
-        if producer is not None:
-            id = producer.op_identifier()
-            if id is not None and id in self._op_to_pattern:
-                return self._op_to_pattern[id]
-        return None
-
-
-class _BacktrackingOr(ValuePattern):
-    """Represents an unrestricted form of OR pattern implemented using backtracking."""
-
-    def __init__(
-        self,
-        values: Sequence[ValuePattern],
-        name: str | None = None,
-        tag_var: str | None = None,
-        tag_values: Sequence[Any] | None = None,
-    ) -> None:
-        """
-        Initialize a _BacktrackingOr pattern.
-
-        Args:
-            values: A sequence of value patterns to match against.
-            name: An optional variable name for the pattern. Defaults to None. If present,
-                this name will be bound to the value matched by the pattern.
-            tag_var: An optional variable name for the tag. Defaults to None. If present,
-                it will be bound to a value (from tag_values) indicating which alternative was matched.
-            tag_values: An optional sequence of values to bind to the tag_var. Defaults to None.
-                If present, the length of tag_values must match the number of alternatives in values.
-                In a successful match, tag-var will be bound to the i-th value in tag_values if the i-th
-                alternative pattern matched. If omitted, the default value of (0, 1, 2, ...) will be used.
-        """
-        super().__init__(name)
-        if tag_values is not None:
-            if tag_var is None:
-                raise ValueError("tag_var must be specified if tag_values is provided.")
-            if len(tag_values) != len(values):
-                raise ValueError(
-                    "tag_values must have the same length as the number of alternatives."
-                )
-        else:
-            tag_values = tuple(range(len(values)))
-        self._tag_var = tag_var
-        self._tag_values = tag_values
-        self._values = values
-
-    @property
-    def tag_var(self) -> str | None:
-        """Returns the tag variable associated with the OrValue pattern."""
-        return self._tag_var
-
-    def clone(self, node_map: dict[NodePattern, NodePattern]) -> _BacktrackingOr:
-        return _BacktrackingOr(
-            [v.clone(node_map) for v in self._values],
-            self.name,
-            self._tag_var,
-            self._tag_values,
-        )
-
-
-def OrValue(
-    values: Sequence[ValuePattern],
-    name: str | None = None,
-    tag_var: str | None = None,
-    tag_values: Sequence[Any] | None = None,
-) -> ValuePattern:
-    """
-    Creates an OR pattern.
-
-    Args:
-        values: A sequence of value patterns to match against.
-        name: An optional variable name for the pattern. Defaults to None. If present,
-            this name will be bound to the value matched by the pattern.
-        tag_var: An optional variable name for the tag. Defaults to None. If present,
-            it will be bound to a value (from tag_values) indicating which alternative was matched.
-        tag_values: An optional sequence of values to bind to the tag_var. Defaults to None.
-            If present, the length of tag_values must match the number of alternatives in values.
-            In a successful match, tag-var will be bound to the i-th value in tag_values if the i-th
-            alternative pattern matched. If omitted, the default value of (0, 1, 2, ...) will be used.
-    """
-    if tag_values is not None:
-        if tag_var is None:
-            raise ValueError("tag_var must be specified if tag_values is provided.")
-        if len(tag_values) != len(values):
-            raise ValueError(
-                "tag_values must have the same length as the number of alternatives."
-            )
-    else:
-        tag_values = tuple(range(len(values)))
-
-    def make_op_id_or_pattern() -> _OpIdDispatchOr | None:
-        mapping: dict[ir.OperatorIdentifier, tuple[Any, NodeOutputPattern]] = {}
-        for i, alternative in enumerate(values):
-            if not isinstance(alternative, NodeOutputPattern):
-                return None
-            producer = alternative.producer()
-            id = producer.op_identifier()
-            if id is None or id in mapping:
-                return None
-            mapping[id] = (tag_values[i], alternative)
-        return _OpIdDispatchOr(mapping, name, tag_var)
-
-    optimized_pattern = make_op_id_or_pattern()
-    return optimized_pattern or _BacktrackingOr(
-        values, name, tag_var, tag_values if tag_var else None
-    )
-
-
-def _nodes_in_pattern(outputs: Sequence[ValuePattern]) -> list[NodePattern]:
-    """Returns all nodes used in a pattern, given the outputs of the pattern."""
-    node_patterns: list[NodePattern] = []
-
-    def visit(value_patterns: Sequence[ValuePattern | None]) -> None:
-        for value_pattern in value_patterns:
-            if isinstance(value_pattern, NodeOutputPattern):
-                node_pattern = value_pattern.producer()
-                if node_pattern not in node_patterns:
-                    node_patterns.append(node_pattern)
-                    visit(node_pattern.inputs)
-
-    visit(outputs)
-    node_patterns.reverse()
-    return node_patterns
-
-
-def _add_backward_slice(
-    node: NodePattern,
-    backward_slice: set[NodePattern],
-    backward_slice_values: set[ValuePattern],
-) -> None:
-    """Adds all nodes in the backward slice of given node to the set `backward_slice`.
-
-    The backward slice of a node is the set of all nodes that are reachable from the node
-    in a backward traversal from the given node.
-    """
-    if node in backward_slice:
-        return
-    backward_slice.add(node)
-    for value_pattern in node.inputs:
-        if isinstance(value_pattern, NodeOutputPattern):
-            _add_backward_slice(
-                value_pattern.producer(), backward_slice, backward_slice_values
-            )
-        elif isinstance(value_pattern, (_OpIdDispatchOr, _BacktrackingOr)):
-            backward_slice_values.add(value_pattern)
-
-
-class GraphPattern:
-    """Represents a pattern that can be matched against a subgraph."""
-
-    def __init__(
-        self,
-        inputs: Sequence[ValuePattern],
-        outputs: Sequence[ValuePattern],
-        nodes: Sequence[NodePattern],
-    ) -> None:
-        self._inputs = inputs
-        self._outputs = outputs
-        if len(outputs) == 0:
-            raise ValueError("GraphPattern must have at least one output")
-        self._nodes = nodes  # _nodes_in_pattern(outputs)
-
-        # Determine the output nodes of the pattern. These are a minimal set of nodes
-        # whose backward-slices cover the entire pattern.
-        output_nodes: set[NodePattern] = set()
-        covered: set[NodePattern] = set()
-        choice_values_returned: set[ValuePattern] = set()
-        covered_choice_values: set[ValuePattern] = set()
-        for value_pattern in outputs:
-            if not isinstance(value_pattern, ValuePattern):
-                raise TypeError(
-                    f"Invalid type {type(value_pattern)} for graph pattern output."
-                )
-            if isinstance(value_pattern, NodeOutputPattern):
-                candidate = value_pattern.producer()
-                if candidate not in covered:
-                    output_nodes.add(candidate)
-                    _add_backward_slice(candidate, covered, covered_choice_values)
-            elif isinstance(value_pattern, (_OpIdDispatchOr, _BacktrackingOr)):
-                choice_values_returned.add(value_pattern)
-
-        # check if all choice_values_returned are contained in covered_choice_values:
-        # We don't yet support the use of a choice-value as a "root" of the search.
-        # This is a limitation of the current implementation, and will be fixed in the future.
-        if not (choice_values_returned <= covered_choice_values):
-            raise NotImplementedError("Returning uncovered choice-values is not supported.")
-
-        self.output_nodes: list[NodePattern] = list(output_nodes)
-
-    @property
-    def output_node(self) -> NodePattern:
-        if len(self.output_nodes) != 1:
-            raise ValueError("GraphPattern does not have unique output node.")
-        return self.output_nodes[0]
-
-    def node(self, index: int) -> NodePattern:
-        return self._nodes[index]
-
-    def num_nodes(self) -> int:
-        return len(self._nodes)
-
-    def __len__(self) -> int:
-        return self.num_nodes()
-
-    @property
-    def inputs(self) -> Sequence[ValuePattern]:
-        return self._inputs
-
-    @property
-    def outputs(self) -> Sequence[ValuePattern]:
-        return self._outputs
-
-    def __iter__(self) -> Iterator[NodePattern]:
-        return iter(self._nodes)
-
-    def __reversed__(self) -> Iterator[NodePattern]:
-        return reversed(self._nodes)
-
-    @property
-    def has_single_output_node(self) -> bool:
-        return len(self.output_nodes) == 1
-
-    @property
-    def num_outputs(self) -> int:
-        return len(self._outputs)
-
-    def commute(self) -> Sequence[GraphPattern]:
-        def commute_node(node: NodePattern) -> Iterable[bool]:
-            if node.op_identifier() == ("", "Add", "") or node.op_identifier() == (
-                "",
-                "Mul",
-                "",
-            ):
-                # Try with and without swapping inputs.
-                return [False, True]
-            # No swapping of inputs
-            return [False]
-
-        iteration_space = [commute_node(node) for node in self._nodes]
-
-        def copy_graph(swap_list: Iterable[bool]) -> GraphPattern:
-            if not any(swap_list):
-                # No need to swap inputs of any node
-                return self
-            # Create a copy of the graph, with swapped inputs for the nodes that need it.
-            node_map: dict[NodePattern, NodePattern] = {}
-            new_inputs = [v.clone(node_map) for v in self._inputs]
-            new_nodes = [
-                node.clone(node_map, swap) for node, swap in zip(self._nodes, swap_list)
-            ]
-            new_outputs = [v.clone(node_map) for v in self._outputs]
-            return GraphPattern(new_inputs, new_outputs, new_nodes)
-
-        return [copy_graph(swap_list) for swap_list in itertools.product(*iteration_space)]
-
-    def __str__(self) -> str:
-        inputs = ", ".join(str(v) for v in self._inputs)
-        outputs = ", ".join(str(v) for v in self._outputs)
-        nodes = "\n   ".join(str(n) for n in self._nodes)
-        return f"pattern ({inputs}) {{\n   {nodes}\n   return {outputs}\n}}"
-
-
-def _to_graph_pattern(pattern_constructor: Callable) -> GraphPattern:
-    """Convert a pattern-construction function to a GraphPattern.
-
-    A pattern-construction function will return values as below:
-    ::
-        def pattern(op, x: Var, shape1: Var, shape2: Var):
-            ...
-            return outputs
-
-    We create a pattern graph by creating pattern-variables for each parameter of the function,
-    and calling the function. The returned values are normalized to a list of ValuePatterns,
-    which represent the outputs of the pattern graph.
-
-    Args:
-        pattern_constructor: Callable
-
-    Returns:
-        GraphPattern: A representation of the pattern that can be matched against a subgraph.
-    """
-    _pattern_vars = inspect.signature(pattern_constructor).parameters
-    pattern_inputs = [Var(v) for v in _pattern_vars][1:]  # Skip the first parameter
-    builder = OpsetPatternBuilder("", record=True)
-    with pattern_builder(builder):
-        pattern_outputs = pattern_constructor(builder, *pattern_inputs)
-    # TODO(rama): classify inputs as value/attribute vars
-    # Returned value could be a single ValuePattern or a list of ValuePatterns.
-    # Normalize representation to a list of ValuePatterns.
-    if isinstance(pattern_outputs, ValuePattern):
-        pattern_outputs = [pattern_outputs]
-    return GraphPattern(pattern_inputs, pattern_outputs, builder.nodes())
-
-
-def _valid_to_replace(
-    matched_nodes: Sequence[ir.Node], output_values: Sequence[ir.Value]
-) -> bool:
-    """Check that values computed by the matched_nodes, except for output_values, are used only by the matched_nodes."""
-    # * Must check that all values matched by pattern are used only by pattern,
-    # except for the value that is replaced.
-    # * Must ensure that replacement subgraph does not use any of the deleted
-    # (intermediate) values. (Not necessary for now. Guaranteed.)
-    for n in matched_nodes:
-        for v in n.outputs:
-            if v in output_values:
-                continue
-            if v.is_graph_output():
-                # value is an output-value of the graph/function.
-                return False
-            for consumer, _ in v.uses():
-                if consumer not in matched_nodes:
-                    return False
-    return True
-
+from onnxscript.rewriter._rewrite_rule import (
+    RewriteRule,
+    RewriteRuleAsClass,
+    RewriteRuleClassBase,
+    RewriteRuleSet,
+    make_rewrite_rule_from_class,
+)
 
 RewriterContext = _tape.Builder
 
-
-@dataclasses.dataclass
-class ReplacementSubgraph:
-    """A subgraph that will replace the matched pattern."""
-
-    match: MatchResult
-    new_outputs: Sequence[ir.Value]
-    new_nodes: Sequence[ir.Node]
-    new_initializers: Sequence[ir.Value]
-    used_opsets: _tape.UsedOpsets
-
-
-def always_true(*args, **kwargs) -> bool:
-    """A condition function that always returns True.
-
-    This is used when no condition function is provided for a rewrite rule.
-    """
-    return True
-
-
-class ReplacementPatternFunction:
-    """The replacement pattern that will replace the targeted pattern.
-
-    Attributes:
-        function (Callable): The replacement function that will be used to replace the matched pattern.
-    """
-
-    def __init__(self, function) -> None:
-        self._function = function
-
-    def get_replacement(self, match: MatchResult) -> ReplacementSubgraph | None:
-        context = RewriterContext()
-        new_outputs = self._function(context, **match.bindings)
-        if new_outputs is None:
-            return None  # Failed to create replacement subgraph
-        if not isinstance(new_outputs, Sequence):
-            new_outputs = [new_outputs]
-        return ReplacementSubgraph(
-            match, new_outputs, context.nodes, context.initializers, context.used_opsets
-        )
-
-
-def _update_opset_imports(
-    graph_or_function: ir.Graph | ir.Function, delta: ReplacementSubgraph
-):
-    imports = graph_or_function.opset_imports
-    for domain, version in delta.used_opsets:
-        if domain not in imports:
-            # use 1 as default version if not explicitly specified
-            imports[domain] = version if version is not None else 1
-        elif version is not None and version != imports[domain]:
-            raise ValueError(
-                f"Multiple versions of opset {domain} used. "
-                f"Expected version {imports[domain]}, but got {version}."
-            )
-
-
-class PatternMatcher(abc.ABC):
-    def __init__(self, pattern: GraphPattern) -> None:
-        self.pattern = pattern
-
-    @abc.abstractmethod
-    def match(
-        self,
-        model: ir.Model,
-        graph_or_function: ir.Graph | ir.Function,
-        node: ir.Node,
-        *,
-        verbose: int = 0,
-        remove_nodes: bool = True,
-        tracer: MatchingTracer | None = None,
-    ) -> MatchResult:
-        """Match the pattern against the subgraph ending at the given node."""
-
-    def __str__(self) -> str:
-        return str(self.pattern)
-
-
-class SimplePatternMatcher(PatternMatcher):
-    def __init__(self, pattern: GraphPattern) -> None:
-        super().__init__(pattern)
-        self._current_node: ir.Node | None = None
-
-    def fail(self, reason: str, node: ir.Node | None = None) -> bool:
-        if self._verbose:
-            num_matched_nodes = self._match.num_matched_nodes()
-            if num_matched_nodes > 0:  # Print only if at least one node successfully matched.
-                print(f"Match failed after {num_matched_nodes} nodes: {reason}")
-        self._match.fail(reason, node or self._current_node)
-        return False
-
-    def _match_constant(self, pattern_constant: Constant, value: ir.Value) -> bool:
-        """Match a Constant pattern against a value.
-
-        If the constant value is produced by a Constant node, we do not include
-        the constant node as part of the matched graph. Thus, it will not be deleted,
-        if subgraph replacement happens. But subsequent DCE will remove the constant
-        node if it is not used elsewhere.
-        """
-        constant_value = value.const_value
-        if constant_value is None:
-            return self.fail(
-                f"Value {value.name} is not a constant, expecting {pattern_constant.value}.",
-            )
-
-        try:
-            constant_value_numpy = constant_value.numpy()
-        except FileNotFoundError:
-            return self.fail(f"Constant value of {value.name} not available.")
-
-        pattern_constant_value = pattern_constant._value
-
-        if isinstance(pattern_constant_value, list):
-            expected_shape = (len(pattern_constant_value),)
-            if constant_value_numpy.shape != expected_shape:
-                return self.fail(f"Value has mismatching shape, expecting {expected_shape}.")
-            if not all(
-                math.isclose(
-                    constant_value_numpy.item(i),
-                    pattern_constant_value[i],
-                    rel_tol=pattern_constant._rel_tol,
-                    abs_tol=pattern_constant._abs_tol,
-                )
-                for i in range(len(pattern_constant_value))
-            ):
-                return self.fail(
-                    f"Value mismatch: expected {pattern_constant_value}, got {constant_value_numpy}."
-                )
-            return True
-
-        # TODO (rama): allow users to specify shape requirement, if desired.
-        if constant_value_numpy.size != 1:
-            return self.fail(
-                f"Value {value.name} is not a scalar, expecting {pattern_constant_value}.",
-            )
-
-        if not math.isclose(
-            constant_value_numpy.item(),
-            pattern_constant_value,
-            rel_tol=pattern_constant._rel_tol,
-            abs_tol=pattern_constant._abs_tol,
-        ):
-            return self.fail(
-                f"Constant value mismatch: expected {pattern_constant_value}, got {constant_value_numpy.item()}.",
-            )
-
-        return True
-
-    def _match_node(self, pattern_node: NodePattern, node: ir.Node) -> bool:
-        """Matches a pattern subgraph against subgraph rooted at node."""
-        self._current_node = node
-        # Graph-matching: we do not allow the same pattern node to be matched against
-        # different graph nodes.
-        matched_node = self._match.lookup_node(pattern_node)
-        if matched_node is not None:
-            if matched_node is not node:
-                return self.fail("Same pattern node is matched against different graph nodes.")
-            return True
-        match = self._match
-        if not pattern_node.matches(node, match):
-            return self.fail(match.reason)
-
-        if self._verbose:
-            print(f"Matched: {node.op_type}")
-
-        match.bind_node(pattern_node, node)
-
-        # TODO: Revisit this to handle optional trailing inputs better.
-        if pattern_node.allow_other_inputs:
-            if len(node.inputs) < len(pattern_node.inputs):
-                return self.fail(
-                    f"Number of inputs ({len(node.inputs)}) is less than expected ({len(pattern_node.inputs)})"
-                )
-        else:
-            if len(node.inputs) != len(pattern_node.inputs):
-                return self.fail(
-                    f"Input nums mismatch. {len(node.inputs)} vs {len(pattern_node.inputs)}"
-                )
-
-        for arg_value, arg_pattern in zip(node.inputs, pattern_node.inputs):
-            # arg_pattern could be a Var, if it's the original arg.
-            if arg_pattern is None:
-                if arg_value is None:
-                    continue
-                else:
-                    return self.fail("(Optional) input is expected to be None but is not.")
-            if not self._match_value(arg_pattern, arg_value):
-                return False
-
-        for i, output_value_pattern in enumerate(pattern_node.outputs):
-            if not self._match.bind_value(output_value_pattern, node.outputs[i]):
-                return False
-
-        return True
-
-    def _match_value(self, pattern_value: ValuePattern, value: ir.Value | None) -> bool:
-        """Match an IR value against a ValuePattern instance."""
-        if isinstance(pattern_value, AnyValue):
-            return True
-
-        if not self._match.bind_value(pattern_value, value):
-            return False
-
-        if isinstance(pattern_value, NodeOutputPattern):
-            if value is None:
-                return self.fail("Mismatch: Computed node pattern does not match None.")
-            return self._match_node_output(pattern_value, value)
-        if isinstance(pattern_value, Constant):
-            if value is None:
-                return self.fail("Mismatch: Constant pattern does not match None.")
-            return self._match_constant(pattern_value, value)
-        if isinstance(pattern_value, _BacktrackingOr):
-            for i, pattern_choice in enumerate(pattern_value._values):
-                self._match.enter_new_match()
-                if self._match_value(pattern_choice, value):
-                    if pattern_value.tag_var is not None:
-                        self._match.bind(pattern_value.tag_var, pattern_value._tag_values[i])
-                    self._match.merge_current_match()
-                    return True
-                self._match.abandon_current_match()
-            return self.fail("None of the alternatives matched.")
-        if isinstance(pattern_value, _OpIdDispatchOr):
-            if value is None:
-                return self.fail("Mismatch: OrValue pattern does not match None.")
-            alternative = pattern_value.get_pattern(value)
-            if alternative is None:
-                return self.fail("Mismatch: OrValue pattern does not match value.")
-            i, pattern_choice = alternative
-            result = self._match_value(pattern_choice, value)
-            if result:
-                if pattern_value.tag_var is not None:
-                    self._match.bind(pattern_value.tag_var, i)
-            return result
-        return True
-
-    def _match_node_output(self, pattern_value: NodeOutputPattern, value: ir.Value) -> bool:
-        """Match an IR value against a NodeOutputPattern instance."""
-        node = value.producer()
-        if node is None:
-            return self.fail(
-                "Mismatch: Computed node pattern does not match uncomputed IR value."
-            )
-        if value.index() != pattern_value.output_index:
-            return self.fail(
-                f"Node output index mismatch: expected {pattern_value._output_index}, got {value.index()}."
-            )
-        return self._match_node(pattern_value.producer(), node)
-
-    def _init_match(self, verbose: int) -> None:
-        """Initialize the match state. Invoked before starting a new match."""
-        self._verbose = verbose
-        self._match: MatchResult = MatchResult()
-        self._current_node = None
-
-    def _get_output_values(self) -> list[ir.Value] | None:
-        """Get values bound to the output variables of the pattern."""
-        output_values: list[ir.Value] = []
-        unbound_values: list[str] = []
-        for j, value_pattern in enumerate(self.pattern.outputs):
-            if value_pattern.name is not None:
-                if value_pattern.name in self._match.bindings:
-                    output_values.append(self._match.bindings[value_pattern.name])
-                else:
-                    unbound_values.append(value_pattern.name)
-            else:
-                if value_pattern in self._match.value_bindings:
-                    output_values.append(self._match.value_bindings[value_pattern])
-                else:
-                    unbound_values.append(f"output_{j}")
-        if unbound_values:
-            self._match.fail(f"Error: Output values not found: {unbound_values}")
-            return None
-        return output_values
-
-    def _match_single_output_node(
-        self,
-        model: ir.Model,
-        graph_or_function: ir.Graph | ir.Function,
-        node: ir.Node,
-        check_removable: bool,
-    ) -> MatchResult:
-        del model
-        del graph_or_function
-
-        pattern = self.pattern
-        match = self._match
-
-        if not pattern.has_single_output_node:
-            return match.fail(
-                "Internal Error: SimplePatternMatcher should not be used for patterns with multiple output nodes."
-            )
-
-        if not self._match_node(pattern.output_node, node):
-            return match
-
-        output_values = self._get_output_values()
-        if output_values is None:
-            # TODO(rama): Is this a valid (useful) case?
-            return match
-        if check_removable and not _valid_to_replace(match.nodes, output_values):
-            # TODO(rama): Match status should be updated to reflect failure reason.
-            return match.fail("Matched nodes have other uses preventing replacement.")
-
-        match.outputs.extend(output_values)
-        return match
-
-    def _multi_match(self, candidate: Iterable[ir.Node], check_removable: bool) -> MatchResult:
-        """Find a match for a pattern with multiple output nodes.
-
-        For a pattern with K output nodes, the input candidate should specify K nodes
-        in the graph that will be matched against the pattern output nodes.
-
-        Args:
-            candidate: An iterable of nodes that will be matched against the pattern output nodes.
-            check_removable: If True, check that the matched nodes can be removed (that is, that
-                they are not used elsewhere in the graph).
-        """
-        match = self._match
-        for pattern_node, node in zip(self.pattern.output_nodes, candidate):
-            if not self._match_node(pattern_node, node):
-                return match
-        output_values = self._get_output_values()
-        if output_values is None:
-            return match
-
-        if check_removable and not _valid_to_replace(match.nodes, output_values):
-            return match.fail("Matched nodes have other uses preventing replacement.")
-
-        match.outputs.extend(output_values)
-        return match
-
-    def match(
-        self,
-        model: ir.Model,
-        graph_or_function: ir.Graph | ir.Function,
-        node: ir.Node,
-        *,
-        verbose: int = 0,
-        remove_nodes: bool = True,
-        tracer: MatchingTracer | None = None,
-    ) -> MatchResult:
-        """Match the pattern against the subgraph ending at the given node.
-
-        For patterns with multiple output nodes, the given node is matched
-        against the first output node in the pattern. For the remaining
-        output nodes in the pattern, we use a brute-force algorithm that
-        enumerates all possible combinations of nodes from the graph (with
-        a filter based on op-type).
-
-        TODO: Consider omitting parameters model and graph_or_function. With
-        the new IR, the graph can be obtained from the node, and the model is
-        not used. But this is a shared abstract method of the Matcher interface,
-        so other matcher implementation also needs to be updated. More importantly,
-        matching in the presence of subgraphs (control-flow) can introduce some
-        complications which require careful consideration.
-        """
-        self._tracer = tracer
-        if self.pattern.has_single_output_node:
-            self._init_match(verbose)
-            return self._match_single_output_node(
-                model, graph_or_function, node, check_removable=remove_nodes
-            )
-        else:
-            # Note: This is a potentially expensive algorithm for matching patterns with
-            # multiple output nodes. For patterns with N output nodes, we try all possible
-            # combinations of N nodes from the graph, and check if they match the pattern.
-            # The first node is fixed to the node argument in this method call. We do
-            # some simple filtering by restricting the candidates for each remaining
-            # output nodes to graph nodes with the same op_type as the corresponding pattern
-            # node. For now, this is intended to be a simple, but robust, implementation
-            # that can be used for debugging and testing. The GenericPatternMatcher is a
-            # more sophisticated implementation, but incomplete.
-            pattern_output_nodes = self.pattern.output_nodes
-            op_to_nodes: dict[tuple[str, str, str], list[ir.Node]] = {}
-            for n in graph_or_function:
-                op_to_nodes.setdefault(n.op_identifier(), []).append(n)
-            all_nodes = iter(graph_or_function)
-
-            def get_nodes(pattern_node):
-                id = pattern_node.op_identifier()
-                if id is None:
-                    return all_nodes
-                return op_to_nodes.get(id, [])
-
-            candidates = [iter([node])] + [get_nodes(pn) for pn in pattern_output_nodes[1:]]
-            match = None
-            for combination in itertools.product(*candidates):
-                self._init_match(verbose)
-                match = self._multi_match(combination, check_removable=remove_nodes)
-                if match:
-                    return match
-            if match is None:
-                return MatchResult().fail("No match found.")
-            return match
-
-
-class RewriteRule:
-    def __init__(
-        self,
-        target_pattern: GraphPattern | Callable,
-        replacement_pattern: ReplacementPatternFunction | Callable,
-        condition_function: Callable | None = None,
-        matcher: PatternMatcher | Callable[[GraphPattern], PatternMatcher] | None = None,
-        verbose: int = 0,
-        name: str | None = None,
-        remove_nodes: bool = True,
-        graph_pre_visitor: Callable[[], None] | None = None,
-        graph_post_visitor: Callable[[], None] | None = None,
-        as_function: bool = False,
-    ) -> None:
-        """Create a rewrite rule.
-
-        Args:
-            target_pattern: The GraphPattern that will be matched against the IR.
-                If a callable is provided, it will be converted to a GraphPattern.
-            replacement_pattern: The ReplacementPatternFunction that will be used to
-                replace the matched pattern. If a callable is provided, it will be
-                converted to a ReplacementPatternFunction.
-            condition_function: The condition function that will be used to check if
-                the pattern match found should be rewritten.
-            matcher: The pattern matcher that will be used to match the pattern.
-                If not provided, a default matcher will be used.
-            verbose: The verbosity level of the rule.
-            name: An optional name for the pattern that will show up in verbose logging.
-            remove_nodes: If True, the matched nodes will be removed from the graph.
-            graph_pre_visitor: A function that will be called before applying the
-                rewriting to the top-level graph or a function.
-            graph_post_visitor: A function that will be called after the rewriting
-                is complete for a graph or function.
-            as_function: If True, the matched nodes will be extracted into a model
-                local function. This is only supported when remove_nodes=True and
-                when the replacement subgraph has a single node, representing the
-                function call.
-        """
-        if as_function and not remove_nodes:
-            raise ValueError("as_function=True is only supported when remove_nodes=True.")
-        if not isinstance(target_pattern, GraphPattern):
-            target_pattern = _to_graph_pattern(target_pattern)
-        self._target_pattern = target_pattern
-
-        if not isinstance(replacement_pattern, ReplacementPatternFunction):
-            replacement_pattern = ReplacementPatternFunction(replacement_pattern)
-        self._replacement_pattern = replacement_pattern
-        self._condition_function = condition_function or always_true
-        if isinstance(matcher, PatternMatcher):
-            self._matcher = matcher
-        elif matcher is None:
-            if target_pattern.has_single_output_node:
-                self._matcher = SimplePatternMatcher(self._target_pattern)
-            else:
-                import onnxscript.rewriter.generic_pattern as generic_pattern
-
-                self._matcher = generic_pattern.GenericPatternMatcher(self._target_pattern)
-        else:
-            self._matcher = matcher(self._target_pattern)
-        self._verbose = verbose
-        self.name = name
-        self.remove_nodes = remove_nodes
-        self.graph_pre_visitor = graph_pre_visitor
-        self.graph_post_visitor = graph_post_visitor
-        self.as_function = as_function
-
-    def __str__(self) -> str:
-        return self.name if self.name else "Anonymous Rule"
-
-    def try_rewrite(
-        self,
-        model: ir.Model,
-        graph_or_function: ir.Graph | ir.Function,
-        node: ir.Node,
-        *,
-        verbose: int | None = None,
-        tracer: MatchingTracer | None = None,
-    ) -> ReplacementSubgraph | None:
-        """If the node matches the pattern, then replace the node with the replacement pattern."""
-        if verbose and verbose > 2:
-            print(f"[try_rewrite] {self}")
-        verbose = verbose if verbose is not None else self._verbose
-        match = self._matcher.match(
-            model, graph_or_function, node, verbose=verbose, remove_nodes=self.remove_nodes
-        )
-        if match:
-            context = None  # TODO(rama)
-            for var in self._target_pattern.inputs:
-                if var.name is not None:
-                    if var.name not in match.bindings:
-                        match.bind(var.name, None)
-            check_match_result = self._condition_function(context, **match.bindings)
-            if not check_match_result:
-                # If check function was provided, but it failed, return the reason for failure to the tracer.
-                if isinstance(check_match_result, MatchResult):
-                    match.fail(
-                        check_match_result.reason,
-                        check_match_result.failure_nodes_and_values,
-                    )
-                if tracer:
-                    tracer.log(
-                        self, graph_or_function, node, match, MatchStatus.CONDITION_FAILED
-                    )
-                return None
-            replacement_subgraph = self._replacement_pattern.get_replacement(match)
-            if replacement_subgraph is None:
-                if tracer:
-                    tracer.log(
-                        self, graph_or_function, node, match, MatchStatus.REPLACEMENT_FAILED
-                    )
-                return None
-            if len(replacement_subgraph.new_outputs) != self._target_pattern.num_outputs:
-                raise ValueError(
-                    f"Number of outputs from replacement function does not match the number of outputs from the target pattern. "
-                    f"Expected {self._target_pattern.num_outputs}, but got {len(replacement_subgraph.new_outputs)}."
-                )
-            # TODO(rama): Remove the opset imports from deleted nodes?
-            _update_opset_imports(graph_or_function, replacement_subgraph)
-            _update_opset_imports(model.graph, replacement_subgraph)
-            if tracer:
-                tracer.log(self, graph_or_function, node, match, MatchStatus.SUCCESS)
-            return replacement_subgraph
-        if tracer:
-            tracer.log(self, graph_or_function, node, match, MatchStatus.NO_MATCH)
-        return None
-
-    def apply_to_model(
-        self,
-        model: ir.Model,
-        *,
-        commute: bool = False,
-        verbose: int | None = None,
-        tracer: MatchingTracer | None = None,
-    ):
-        # A convenience method to apply the rule to a model. We use a RewriteRuleSet to
-        # handle commutative rules.
-        return RewriteRuleSet([self], commute=commute).apply_to_model(
-            model, verbose=verbose, tracer=tracer
-        )
-
-    def commute(self) -> Sequence[RewriteRule]:
-        def replace_pattern(new_pattern):
-            """Return a shallow copy of self with node_pattern replaced by new_pattern."""
-            # TODO(rama): Maybe we should use a better alternative to construct new matcher.
-            matcher_class = type(self._matcher)
-            return RewriteRule(
-                new_pattern,
-                self._replacement_pattern,
-                self._condition_function,
-                matcher_class(new_pattern),
-                self._verbose,
-                self.name,
-                self.remove_nodes,
-                self.graph_pre_visitor,
-                self.graph_post_visitor,
-                self.as_function,
-            )
-
-        return [replace_pattern(p) for p in self._target_pattern.commute()]
-
-
-class RewriteRuleAsClass:
-    """Defines a class grouping method pattern, rewrite, check.
-    This class is then given to function :func:`make_rewrite_rule_from_class`
-    to define a new rule.
-    """
-
-    @classmethod
-    def pattern(cls, op, *_) -> Any:
-        raise NotImplementedError("Method 'pattern' must be overwritten.")
-
-    @classmethod
-    def rewrite(cls, op, *_) -> Any:
-        raise NotImplementedError("Method 'rewrite' must be overwritten.")
-
-    @classmethod
-    def check(cls, context, *_, **__) -> bool | MatchResult:
-        return MatchResult()
-
-
-def make_rewrite_rule_from_class(
-    rule_class: type | RewriteRuleAsClass, generic: bool = False
-) -> RewriteRule:
-    """Creates a RewriteRule from a class defining the function
-    pattern, rewrite, check with class method. It makes it is easier
-    to read when a module contains multiple patterns.
-
-    Example::
-
-        class TransposeIdentity(RewriteRuleAsClass):
-            @classmethod
-            def pattern(cls, op, x, perm):
-                return op.Transpose(x, perm=perm)
-
-            @classmethod
-            def check(cls, context, x: ir.Value, perm: ir.Attr | ir.RefAttr) -> bool:
-                if isinstance(perm, ir.RefAttr):
-                    return False
-                if perm.type == ir.AttributeType.INTS:
-                    if perm.value == list(range(len(perm.value))):
-                        return True
-                return False
-
-            @classmethod
-            def rewrite(cls, op, x: ir.Value, perm: ir.Attr | None = None):
-                return op.Identity(x)
-
-        transpose_identity_rule = make_rewrite_rule_from_class(TransposeIdentity)
-    """
-    assert hasattr(rule_class, "pattern"), f"Method 'pattern' is missing from {rule_class!r}."
-    assert hasattr(rule_class, "rewrite"), f"Method 'rewrite' is missing from {rule_class!r}."
-    assert hasattr(rule_class, "check"), f"Method 'check' is missing from {rule_class!r}."
-    if generic:
-        import onnxscript.rewriter.generic_pattern as orpp
-
-        return RewriteRule(
-            rule_class.pattern,
-            rule_class.rewrite,
-            rule_class.check,
-            orpp.GenericPatternMatcher,
-            name=rule_class.__name__,  # type: ignore[union-attr]
-        )
-    return RewriteRule(
-        rule_class.pattern,
-        rule_class.rewrite,
-        rule_class.check,
-        name=rule_class.__name__,  # type: ignore[union-attr]
-    )
-
-
-# Variation of RewriteRuleAsClass that is based on instance methods instead of class methods.
-# Useful to implement a family of rules to support pattern variations.
-# TODO: cleanup the naming conventions for these inter-related classes.
-class RewriteRuleClassBase:
-    @classmethod
-    def rule(cls, *args, **kwargs):
-        instance = cls(*args, **kwargs)
-        return RewriteRule(
-            instance.pattern,
-            instance.rewrite,
-            instance.check,
-            name=instance.name,
-            remove_nodes=instance.remove_nodes,
-            graph_pre_visitor=instance.setup,
-            graph_post_visitor=instance.cleanup,
-            as_function=instance.as_function,
-        )
-
-    def __init__(
-        self, name: str | None = None, remove_nodes: bool = True, as_function: bool = False
-    ) -> None:
-        self.name = name or self.__class__.__name__
-        self.remove_nodes = remove_nodes
-        self.as_function = as_function
-
-    def pattern(self, op, *args, **kwargs):
-        raise NotImplementedError("Method 'pattern' must be implemented by derived class.")
-
-    def check(self, op, *args, **kwargs):
-        # Default check function that returns a
-        # MatchResult object with success always set to True.
-        return MatchResult()
-
-    def rewrite(self, op, *args, **kwargs):
-        raise NotImplementedError("Method 'rewrite' must be implemented by derived class.")
-
-    def setup(self):
-        # Optional setup function that can be overridden by derived classes. Used to do
-        # per model/function initialization.
-        pass
-
-    def cleanup(self):
-        # Optional cleanup function that can be overridden by derived classes. Used to do
-        # per model/function cleanup.
-        pass
-
-
-def _copy_for_function(
-    inputs: Sequence[ir.Value | None], nodes: Sequence[ir.Node], outputs: Sequence[ir.Value]
-):
-    """Utility function to extract a subgraph out as a function."""
-    value_map: dict[ir.Value, ir.Value] = {}
-    function_inputs: list[ir.Value] = []
-    constant_nodes: list[ir.Node] = []
-    for input in inputs:
-        # Create a function input (formal-parameter value) to represent this value:
-        new_value = (
-            ir.Value(
-                name=input.name,
-                shape=input.shape,
-                type=input.type,
-                doc_string=input.doc_string,
-            )
-            if input
-            else ir.Value()  # dummy parameter for a None input
-        )
-        if input is not None:
-            value_map[input] = new_value
-        function_inputs.append(new_value)
-
-    def copy_value(value: ir.Value | None) -> ir.Value | None:
-        if value is None:
-            return None
-        if value not in value_map:
-            const_value = value.const_value
-            if const_value is not None:
-                # create a Constant node to represent the value
-                value_attr = ir.AttrTensor("value", const_value)
-                const_node = ir.Node("", "Constant", [], [value_attr])
-                constant_nodes.append(const_node)
-                value_map[value] = result = const_node.outputs[0]
-                return result
-            raise ValueError(f"Value {value} not found in value_map.")
-        return value_map[value]
-
-    def copy_attr_value(attr: ir.Attr | ir.RefAttr) -> ir.Attr | ir.RefAttr:
-        if not isinstance(attr, ir.Attr):
-            # No need to support this currently, as rewriting inside a function is
-            # not used, as it has several challenges.
-            raise NotImplementedError("RefAttr not supported.")
-        if attr.type in {ir.AttributeType.GRAPH, ir.AttributeType.GRAPHS}:
-            # No need to support this currently, as rewriting control-flow constructs
-            # is not used and has several challenges.
-            raise NotImplementedError("Graph attributes not supported.")
-        # Primitive attributes are immutable by design and can be shared.
-        return attr
-
-    def copy_node(node: ir.Node) -> ir.Node:
-        new_inputs = [copy_value(v) for v in node.inputs]
-        new_attributes = [copy_attr_value(v) for v in node.attributes.values()]
-        new_node = ir.Node(
-            node.domain,
-            node.op_type,
-            new_inputs,
-            new_attributes,
-            overload=node.overload,
-            num_outputs=len(node.outputs),
-            graph=None,
-            name=node.name,
-            doc_string=node.doc_string,  # type: ignore
-            metadata_props=node.metadata_props.copy(),
-        )
-        new_outputs = new_node.outputs
-        for i, output in enumerate(node.outputs):
-            value_map[output] = new_outputs[i]
-            if output.name is not None:
-                new_outputs[i].name = output.name
-        return new_node
-
-    function_nodes = [copy_node(node) for node in nodes]
-    function_outputs = [copy_value(v) for v in outputs]
-    return (function_inputs, constant_nodes + function_nodes, function_outputs)
-
-
-def _get_new_overload(model: ir.Model, domain: str, name: str) -> str:
-    """Get a new overload for the given domain and name.
-
-    Args:
-        model: The model to which the new overload will be added.
-        domain: The domain of the new overload.
-        name: The opname of the new overload.
-
-    Returns:
-        The new overload name.
-    """
-    existing_functions = model.functions
-    # Just a simple implementation for now
-    overload = 1
-    while True:
-        overload_name = str(overload)
-        if (domain, name, overload_name) not in existing_functions:
-            return overload_name
-        overload += 1
-
-
-class RewriteRuleSet:
-    def __init__(self, rules: Sequence[RewriteRule], *, commute: bool = False) -> None:
-        if not rules:
-            raise ValueError("rules must contain at least one rule")
-        if commute:
-            rules = list(itertools.chain.from_iterable([rule.commute() for rule in rules]))
-        self.rules = rules
-        # We call remove_unused_nodes at end of rewriting if there is any rule that does
-        # NOT remove nodes (immediately when it is applied)
-        self.remove_unused_nodes = any(not rule.remove_nodes for rule in rules)
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}({self.rules})"
-
-    def _apply_to_graph_or_function(
-        self,
-        model: ir.Model,
-        graph_or_function: ir.Graph | ir.Function,
-        *,
-        verbose: int | None,
-        tracer: MatchingTracer | None = None,
-    ) -> int:
-        """
-        Apply the rewrite rules to the given graph or function.
-
-        Args:
-            model: The model to which the rewrite rules are applied.
-            graph_or_function: The graph or function to which the rewrite rules are applied.
-            verbose: The verbosity level. Defaults to None.
-            tracer: The tracer for debugging. Defaults to None.
-
-        Returns:
-            The number of rewrite rules applied.
-        """
-        count = 0
-
-        # NOTE: Rules should be prioritized in the order they are added to the RewriteRuleSet.
-        # And the graph is applied in order.
-        for rule in self.rules:
-            if rule.graph_pre_visitor:
-                rule.graph_pre_visitor()
-            for node in graph_or_function:
-                delta = rule.try_rewrite(
-                    model, graph_or_function, node, verbose=verbose, tracer=tracer
-                )
-                if delta is None or tracer is not None:
-                    continue
-                assert isinstance(delta, ReplacementSubgraph)
-                if delta.new_initializers:
-                    if isinstance(graph_or_function, ir.Function):
-                        # TODO(rama): Can't add initializers to functions. But currently this is not
-                        # an issue, as we apply inlining before applying rewrite rules.
-                        if verbose:
-                            print(
-                                f"Rewrites adding initializers not supported for functions: {rule}"
-                            )
-                        continue
-                    initializers = graph_or_function.initializers
-                    for initializer in delta.new_initializers:
-                        if initializer.name in initializers:
-                            if verbose:
-                                print(f"Initializer {initializer.name} already exists.")
-                            continue
-                    for initializer in delta.new_initializers:
-                        initializers[initializer.name] = initializer  # type: ignore[index]
-                # TODO: This does not yet handle the problem of determining the correct insertion point
-                # for inserted nodes in the case of patterns with multiple output-nodes. The following
-                # is sufficient for patterns with a single output-node "node", which can serve as the
-                # insertion-point.
-                onnxscript.optimizer.basic_constant_propagation(delta.new_nodes)
-                if rule.as_function:
-                    # Create a function out of a copy of the matched nodes
-                    if len(delta.new_nodes) != 1:
-                        raise ValueError(
-                            "as_function=True is only supported for patterns with a single replacement node."
-                        )
-                    call_node = delta.new_nodes[0]
-                    domain = call_node.domain
-                    name = call_node.op_type
-                    overload = _get_new_overload(model, domain, name)
-                    call_node.overload = overload
-
-                    # Create topologically sorted list of nodes to be replaced.
-                    unsorted_nodes = set(delta.match.nodes)
-                    original_nodes = [n for n in graph_or_function if n in unsorted_nodes]
-                    # Create new inputs/nodes/outputs for the function
-                    inputs, nodes, outputs = _copy_for_function(
-                        call_node.inputs, original_nodes, delta.match.outputs
-                    )
-
-                    used_domains: set[str] = {node.domain for node in original_nodes}
-                    parent_opset_imports = graph_or_function.opset_imports
-                    used_opset_imports = {
-                        k: v for k, v in parent_opset_imports.items() if k in used_domains
-                    }
-
-                    graph = ir.Graph(
-                        inputs, outputs, nodes=nodes, opset_imports=used_opset_imports
-                    )
-                    f = ir.Function(domain, name, overload, graph=graph, attributes=())
-                    model.functions[f.identifier()] = f
-                _convenience.replace_nodes_and_values(
-                    graph_or_function,
-                    node,
-                    delta.match.nodes if rule.remove_nodes else [],
-                    delta.new_nodes,
-                    delta.match.outputs,
-                    delta.new_outputs,
-                )
-
-                count += 1
-            if rule.graph_post_visitor:
-                rule.graph_post_visitor()
-
-        return count
-
-    def apply_to_model(
-        self,
-        model: ir.Model,
-        *,
-        verbose: int | None = None,
-        tracer: MatchingTracer | None = None,
-    ) -> int:
-        """Apply the rewrite rules in the set to the model.
-
-        Args:
-            model: The model to which the rewrite rules are applied.
-            verbose: The verbosity level of messages. Defaults to None.
-            tracer: if specified, no changes are made to the model, only
-                information about the best matches found is computed.
-
-        Returns:
-            The number of applications of rewrite rules.
-        """
-        assert isinstance(model, ir.Model)
-        onnxscript.optimizer.basic_constant_propagation(model.graph)
-        # Rewriting may introduce new functions. In the following loop,
-        # we restrict rewriting to original functions, not newly introduced ones.
-        original_functions = list(model.functions.values())
-        count = self._apply_to_graph_or_function(
-            model, model.graph, verbose=verbose, tracer=tracer
-        )
-        for function in original_functions:
-            onnxscript.optimizer.basic_constant_propagation(function)
-            count += self._apply_to_graph_or_function(
-                model, function, verbose=verbose, tracer=tracer
-            )
-        if self.remove_unused_nodes:
-            onnxscript.optimizer.remove_unused_nodes(model)
-        return count
-
-    def __iter__(self):
-        yield from self.rules
-
-
-class MatchStatus(enum.IntEnum):
-    """The status of a pattern-matching operation."""
-
-    NO_MATCH = 0  # No successful match found for entire pattern graph
-    CONDITION_FAILED = 1  # Subsequent validation check failed
-    REPLACEMENT_FAILED = 2  # Replacement subgraph could not be created
-    SUCCESS = 3  # A successful match was found
-
-
-@dataclasses.dataclass
-class MatchInfo:
-    """The status of a pattern-matching operation. An extension of MatchResult."""
-
-    match_result: MatchResult
-    root_node: ir.Node
-    container: ir.Graph | ir.Function
-    status: MatchStatus
-
-    def score(self) -> int:
-        """Return a score for the match."""
-        return len(self.match_result.nodes) + int(self.status.value) * 100
-
-    def print(self):
-        separator = "-" * 80
-        print(separator)
-        print(f"Status: {self.status.name}")
-        if self.status != MatchStatus.SUCCESS:
-            reason = self.match_result.reason
-            if reason:
-                if self.status == MatchStatus.CONDITION_FAILED:
-                    print(f"Graph matching failed due to failing check condition : {reason}")
-                else:
-                    print(f"Graph matching failed: {reason}")
-            else:
-                print("Graph matching failed.")
-            failure_nodes_and_values = self.match_result.failure_nodes_and_values
-            print("Failure at or around nodes/values:")
-            if failure_nodes_and_values:
-                for failure_cause in failure_nodes_and_values:
-                    failure_cause.display()
-        print("Matched nodes:")
-        import onnxscript.rewriter._ir_utils as ir_utils
-
-        ir_utils.display_nodes(self.match_result.nodes)
-        print(separator)
-
-
-class MatchingTracer:
-    """A debugging helper class to trace the matching of a pattern against a graph.
-
-    This is used to track the best matches found for each rule, and to report the
-    results at the end of the matching.
-    """
-
-    def __init__(self) -> None:
-        self._best_matches_map: dict[RewriteRule, list[MatchInfo]] = defaultdict(list)
-
-    @property
-    def best_matches_map(self) -> dict[RewriteRule, list[MatchInfo]]:
-        return self._best_matches_map
-
-    def log(
-        self,
-        rule: RewriteRule,
-        container: ir.Graph | ir.Function,
-        node: ir.Node,
-        match_result: MatchResult,
-        status: MatchStatus,
-    ) -> None:
-        this_match = MatchInfo(match_result, node, container, status)
-        this_score = this_match.score()
-        if this_score == 0:
-            return
-        best_matches = self._best_matches_map[rule]
-        if best_matches:
-            if this_score < best_matches[0].score():
-                return
-            if this_score > best_matches[0].score():
-                best_matches.clear()
-        best_matches.append(this_match)
-
-    def report(self) -> None:
-        best_score = 0
-        for rule, matches in self._best_matches_map.items():
-            if not matches:
-                continue
-            if matches[0].score() > best_score:
-                best_score = matches[0].score()
-                best_match = matches[0]
-                best_rule = rule
-
-        if best_score > 0:
-            print(f"Rule: {best_rule}")
-            best_match.print()
-        else:
-            print("No matches found.")
+__all__ = [
+    "ANY_VALUE",
+    "OrValue",
+    "Constant",
+    "OpsetPatternBuilder",
+    "pattern_builder",
+    "RewriteRule",
+    "RewriteRuleAsClass",
+    "RewriteRuleClassBase",
+    "RewriteRuleSet",
+    "make_rewrite_rule_from_class",
+    "RewriterContext",
+    "MatchingTracer",
+    "MatchResult",
+    "MatchStatus",
+    "PatternMatcher",
+    "SimplePatternMatcher",
+]
